requirements.txt
requests==2.31.0
Pillow==10.2.0
openai==1.12.0
reportlab==4.1.0
python-dotenv==1.0.1 